+++ lib/event_bloc_widgets.dart
@@ -3,12 +3,10 @@
 /// See [RepositoryProvider] and [BlocProvider] for good places to start.
 library event_bloc_widgets;
 
-// export 'package:event_bloc/src/widgets/bloc_provider.dart';
-// export 'package:event_bloc/src/widgets/event_channel.dart';
-// export 'package:event_bloc/src/widgets/repository_provider.dart';
-
 export 'package:event_bloc/event_bloc.dart';
 
 export 'package:event_bloc/src/widget/bloc_provider.dart';
 export 'package:event_bloc/src/widget/context.dart';
+export 'package:event_bloc/src/widget/environment.dart';
+export 'package:event_bloc/src/widget/multi_provider.dart';
 export 'package:event_bloc/src/widget/repository_provider.dart';